adding wait_for_postgres script and reverting DB default prefix length change
