+++ packages/rollup-core/config/env_var_updates.config
@@ -6,11 +6,4 @@
 # set the value to $DELETE$ to clear any set value it may have.
 
 #DUMMY_ENV_VAR=somevalue
-
-# REQUEST_LIMIT_PERIOD_MILLIS=
-# MAX_TRANSACTIONS_PER_UNIT_TIME=
-# MAX_NON_TRANSACTION_REQUESTS_PER_UNIT_TIME=
-
-# CONTRACT_DEPLOYER_ADDRESS=
-# COMMA_SEPARATED_TO_ADDRESS_WHITELIST=
-# COMMA_SEPARATED_RATE_LIMIT_WHITELISTED_IPS=
+#VAL_TO_BE_DELETED=$DELETE$
