+++ packages/rollup-services/exec/wait_for_postgres.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+RETRIES=30
+
+until psql -h $POSTGRES_HOST -p $POSTGRES_PORT -U $POSTGRES_USER -d $POSTGRES_DATABASE -c "select 1" > /dev/null 2>&1 || [ $RETRIES -eq 0 ]; do
+  echo "Waiting for postgres server, $((RETRIES--)) remaining attempts..."
+  sleep 1
+done
+
+if [ $RETRIES -eq 0 ]; then
+  echo "Timeout reached waiting for postgres!"
+  exit 1
+fi
+
+echo "Connected to postgres. Continuing with startup..."
